+++ header/message.h
@@ -48,7 +48,7 @@ struct message{
     uint16_t dest_port;
     uint16_t column; 
     uint16_t flag; // 1 accept 0 deny
-    uint32_t nonce;
+    uint64_t nonce;
     char* peerkey;
     uint16_t pkey_len;
     unsigned char* sign;
