+++ header/receive.h
@@ -40,4 +40,4 @@ void changeKeyClientReciver(unsigned char *newKey, int size);
 
 int deserialize_message(unsigned char* buffer, struct message *aux, uint8_t isEncr);
 
-int recv_message(int socket, struct message* message, struct sockaddr* mitt_addr, int dec, uint32_t nonce);
+int recv_message(int socket, struct message* message, struct sockaddr* mitt_addr, int dec, uint64_t nonce);
